+++ lib/devup/env.rb
@@ -1,0 +1,10 @@
+require "devup/version"
+require "devup/environment"
+require "devup/application"
+
+module Devup
+  class Error < StandardError; end
+end
+
+app = Devup::Application.new
+app.load_env